Build the main API Components
The Components that i've built is related to registration
system and account management, now users can register on the system,
and manage their accounts with no problems (hopefully). There's some
components that are not related to the main purpose of this
commit (building registration system), but it will be related in somehow
with others.