Fix "Swagger" issue while generating documentation
QuerySet is not defined inside "ListVehicleStations" view, which cause a
problem while generating the documentation.

The queryset of getting all stations that a specific vehicle belongs to
is written manually, without specifying a "queryset" property inside the
"vehicles_manager.views.ListVehicleStations", this thing causes a
problem with "drf-yasg" (responsible on generating API doc). To solve
it,define view's queryset as an empty queryset.