Build e-wallet application inside the system
Our system has an embedded e-wallet, which is responsible on managing
all payment operations inside the system.

At this time, the e-wallet supports the fundimental functionalities, there's too many things to be
improved, and i am sure that there's to many bugs to be discovered and fixed, But for now,
the e-wallet can do basic operations with no problems.