Build Recommendation System
This system is specified for shortest path suggestion, i've build this
system. This system is just for trying the classification model that
i've used.

The network graph is a random graph from network repository website,
before using this model we need to find the embeddings for each node
inside of the network, then normalize it, and pass selected nodes'
embeddings for the classification model(SVM model).